Add files via upload
This is a network diagram that shows my entire cloud setup, including my Ansible jump box and the Docker containers running on each VM.  I created this network infrastructure in Azure. The goal was to create a highly available web server for  "XCorp's Red Team" to use for testing and training. This was a homework challenge from my UPENN cyber security Bootcamp.